Change data format
For this file, I added a title row and then I replaced all of the commas in the data set with spaces because when I would run it through R, the names were showing up with a period attached to them. For some reason when I try to read the csv file to the "adult" it tells me there are more columns than column names and I can't figure out how to remove all of the blank columns so if one of you wants to do that and upload a new csv file we can code with that.