Add files via upload
I did all three types of the step function with the subsections for the difference in AIC and BIC outputs. I tried to use the regsubsets() function to search for all possible models, but I kept getting an unusual output. I'm going to email her and ask why I might be getting that output.